meta-isar: deploy-image: Change reboot logic
Pull the reboot out of the script. This allows for cleaner integration
with different calling environment, may they be a systemd unit, an
initramfs script or simply a shell for testing purposes.

And if the script exits with an error, wait a minute before rebooting
the system, rather than just trying to re-execute it. This permits to
inspect potential error as well.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>