meta-isar: deploy-image: Drop umount attempt after installation
Under systemd, this fails because /install is still busy. In that setup,
systemd will take care of properly unmounting this partition anyway on
reboot. In other setups, the caller should take care.

This also obsoletes the "set -e" because no commands are left that can
produce errors and are not checked explicitly.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>