meta-isar: deploy-image: Fix and enhance image selection dialog
The dialog was broken because an array of tag/value tuples had to be
provided, not just a list of files.

Use the chance an also fall back to the dialog if the search for
DISK_IMAGE does deliver a result but not a clear one. But only open the
dialog if there are entries.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>