meta-isar: Added support for nfs based rootfs in deploy-image-wic (isar-image-installer).
When AUTO_INSTALL is not used, the script tries to resolve possible target devices to
install the wic image to. To not overwrite the device hosting the current root
we are using, it tries to detect the current root. Which works fine when a local device
is used but does not work in the case of nfs. (most probably it won't work for cifs as well)

With this patch we are detecting nfs based mounts of / before the local device tailored handling
is applied and bypass the breaking parts (in this particular case readlink).

Thus, in case of an nfs the current root device will be "nfs".

Signed-off-by: Alexander Heinisch <alexander.heinisch@siemens.com>