repository: repo_del_package should not remove source packages
When the architecture of the binary package is "all", repo_del_package would
remove the architecture filter to remove binary for any packages. This
unfortunately causes reprepro to also remove the source package. Modify this
function to use "reprepro removefilter" where we can make sure that only .deb
packages are removed.

Signed-off-by: Cedric Hombourger <cedric.hombourger@siemens.com>