remove trailing slash on source list entries
The deb and deb-src URLs in the official debian sources lists do not end
with a slash. For unknown reasons this was not true for some of our
list entries. While this does not make any problems functionally, it has
an effect on caching, as our accesses cannot be cache-shared with the
ones from other users (like docker container builds). It also leads to
ugly looking access path when using snapshot mirrors. We fix it by
removing the trailing slash.

No functional change.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>