linux-custom: support to add linux-libc-dev package with kernel name
Debian and Ubuntu use unified kernel headers and a single linux-libc-dev
package is shipped. In contrast, Isar may build multiple kernel variants
with different versions and/or patches and their packages but linux-libc-dev
are suffixed with -${KERNEL_NAME_PROVIDED}: this allows
each variant to exist in isar-apt (reprepro may not have more than one
version of a given package for a given suite).
Apply the same suffix to linux-libc-dev / linux-libc-dev-${DISTRO_ARCH}-cross
and a Provides: clause.

Signed-off-by: Arulpandiyan Vadivel <arulpandiyan.vadivel@siemens.com>
Signed-off-by: Arulpandiyan Vadivel <arulpandiyan.vadivel@siemens.com>