linux-custom: Pin buster compat level to 10 to fix kernel builds
Compat level 12 of buster selects dh_dwz which breaks in that old
version with certain kernels, e.g. like this:

 dwz: dwz.c:9255: recompute_abbrevs: Assertion `off == cu_size' failed.
 dh_dwz: dwz -q -- debian/linux-image-cip-dbg/usr/lib/debug/lib/modules/5.10.233-cip56/vmlinux died with signal 6

As buster is LTS now, we take the low-risk path and switch back to the
compat level the recipe was once validated again, rather than trying to
resolve this and possibly more issues.

Fixes: ca786c36bca9 ("meta/recipes-kernel: use debhelper-compat to generate compat file")
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>