debianize: make default changelog timestamp configurable
Add the `DEBIAN_CHANGELOG_TIMESTAMP` variable to make the default
changelog timestamp configurable. This might be required if the default
timestamp leads to problems in the build process.

This is the case for e.g. python packages that use distutils<3.8 or with
the hatchling build system, since the internally used zip does not
support timestamps before 1. January 1980.

Signed-off-by: Christoph Steiger <christoph.steiger@siemens.com>