container-loader: make generated package arch specific
The generated package contains files which are architecture specific
(the container image). By that, it must not be arch=all. Further, we
currently cannot co-install the package for multiple architectures and
we also do not want to automatically allow resolving this package in a
foreign architecture. Hence set the multi-arch specifier to "allowed",
which means the dependency needs to be annotated with ":any".

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>