testsuite: Disable "snapshot" feature for installer image.
In the testsuite we use qemu-system with "-shapshot" optiop with wic
images thus preventing them from unwanted modification by tests.

For installer.wic it's reasonable to disable this feature so that
bmaptool will write directly to "installer.wic" instead of temporary
file used by qemu for providing snapshot features.

It also allows to mount it after build for debugging purposes.