isar-apt: support further optional fields in Release
The isar-apt repository is a proper apt repository which can be
distributed as-is. A common use-case for that is to build packages which
are then used in a standard Debian by adding the repository to the apt
sources. In this case, it is desired to fill some of the optional fields
in Release to enable fine-grained pinnings.

We extend the repo_create function to allow appending arbitrary fields
to the Release file. To make that easy to use, we introduce the
ISAR_APT_OPT_FIELD bitbake var map. As the do_cache_config task now
depends on user-input, we also ensure it is rebuild when changed.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>