isar-sstate: add filter parameter
Currently all sstate operations are always performend on the whole
cache. This is problematic if some files should not be uploaded or
a clean operation should only be applied to some files.

We now add the filter parameter, which can be used in combination with
any isar-sstate command. By that, the operation is limited to the files
matching the filter. Negative matches are also supported.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>