add efi booting support for rpi-arm64-v8
Booting the RPI via EFI makes it much easier to decouple the image
contents, firmware, boot config and kernel cmdline. For that, we add a
new target rpi-arm64-v8-efi that boots the RPI using u-boot efi. We
further switch that target to debian bookworm (instead of raspios
bookworm) and by that have a baseline that is closer to upstream.

The recipes are written in a way to make it easy to customize in
downstream layers and only carry a bare minimum of configuration.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>