Pull buster from archive.d.o
Debian buster is now out of LTS, so update the URLs to continue allowing
to build against it. Security updates are no longer provided this way.
For supported packages, those are now coming from Extended LTS [1]. But
as the provider is asking to not use their server heavily, we refrain
from providing those sources here.

[1] https://www.freexian.com/lts/extended/docs/how-to-use-extended-lts/

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>